Awesome toolchain update
List of changes
# Directory structure is now broken up into src, obj, and bin
directories in order to reduce clutter. .asm files should also be very
easy to compile in as well; just add them to the src directory exactly
like you would a .c file
# Startup module reduces normal flash wait states by 2 because of the
use of bootcode C helper functions
# Startup module uses SMC to save/restore items rather than using the
data segment
# math.h -- added a bunch of predefines
# Began framework for an interrupt-driven library
# Changed default icon used for demo programs
# Inclusion of the icon is now completely automatic; if 'iconc.png'
exists in the root of the project directory, it is automatically built
# Updated versions of ConvHEX and ConvPNG were added to the /bin
directory
# Added support to the makefiles for 11 possible shared libraries -- as
of right now there are 3. If someone can find a better solution to my
solution, that would be much appreciated
# Make should now be able to support the -jX flag
# dbg_printf and dbg_sprintf are now defined as the same things;
dbg_sprintf is preferred I guess
# Other random fixes here and there :)