makefile: CCDEBUG: disable -gdwarf-5 -g3 (produce no debug info).
Due to an issue with ez80-clang 15.0.7, this is disabled for now.

Not that this makes no difference for users, as there exists no tool
that uses the produced .dbg files.
CEmu's source-level-debugging branch did use them, but was never finished
and does not actually work in its current state.